att - ad_conversion

+++ include/ad_conversion.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 #include <avr/interrupt.h>
 
-uint16_t ad_read[6];
+uint8_t ad_read[6];
 
 void set_ad_chanel_service(unsigned char channel);
 uint8_t ad_read_service(void);
